어댑터 정리
